+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 1.  Add `Window/Saints/Draw` menu to toggle gizmos drawing. Useful when you want to do some profiler but `gizmos` drawing is too heavy.
 2.  Gizmos drawing is disable in build (it will not draw anything in build anyway)
+3.  Add `DrawWireRectTransform` for `RectTransform` in UI